requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.11.0
django-on-heroku==1.1.2
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
